Extract `getAvailableXAxisScales`
